Sending ADC data to phone over bluetooth.
Modified Hello_World, which was the Bluno demo, to include the information for the ADC chip. Used the information learned in previous experiment to get raw ADC values from the HX711. This was then connected to the example Bluno app. Work was done to get the data in the right format and changing raw ADC values were seen on the phone and the serial terminal.